+++ NEWS.md
@@ -4,6 +4,14 @@ This page describes the most important changes in `AstroFITS`. The format is bas
 Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic
 Versioning](https://semver.org).
 
+## Unreleased
+
+### Fixed
+
+- Key can be symbolic in `Base.get(T, hdu, key[, def])`.
+
+- Non-existing symbol `write!` is no longer exported.
+
 ## Version 1.1.0 [2026-01-22]
 
 ### Added